+++ test/ApplicationData/ApplicationDataTests.vcxproj.filters
@@ -21,6 +21,9 @@
     <ClCompile Include="ApplicationDataTests.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="ApplicationDataTests_Elevated.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="pch.h">
